[fee_handler] rename `WithdrawFee` to `WithdrawCredit` and move it to its own module.
