+++ primitives/fee-handler/src/lib.rs
@@ -1,3 +1,4 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 pub mod fee_handler;
+pub mod withdraw_credit;
